requirements_openai.txt
streamlit>=1.29.0
openai>=1.0.0
chromadb>=0.4.22
pypdf2>=3.0.1
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
tiktoken>=0.5.2
numpy>=1.24.3
pandas>=2.0.0
